Use bigger format for object id render texture to store full entity id

+++ hyperion/include/hyperion/render/frame/context/render_frame_context_objects.hpp
@@ -10,7 +10,7 @@
 //-------------------- Definition Namespace --------------------
 namespace Hyperion::Rendering {
 
-  using RenderFrameContextObjectId = uint32;
+  using RenderFrameContextObjectId = uint64;
 
   struct RenderFrameContextObject {
     RenderFrameContextObjectId id = 0;
